requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.1.0+cpu
torchvision==0.16.0+cpu
Flask==2.3.3
Pillow==10.0.1
numpy<2